+++ Fun.Blazor.Server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+## [3.2.3] - 2023-11-22
+
+- Add MapBlazorSSRComponents
+- Add MapFunBlazorCustomElements
+
 ## [3.2.0] - 2023-11-16
 
 - Unify version